makes max runtime setting a bit softer
Right now, we do not schedule new jobs as soon as the max runtime has been
reached and wait for the running ones to terminate. In the unfortunate
situation that there is a job running that might take a bit longer, this will
cause new jobs not to be processed for some time.

With this change, we wait for running jobs to terminate for some time. If
they haven't terminated until then, we will start new jobs again, and
repeat.